feat(integrator): added euler method for integration of orientation #1012
Added the Euler method for integrating the orientation of the IMU.
Weirdly, this simpler and theoretically worse algorithm is outputting
errors much smaller than RK4. A bug in the implementation of RK4 is
probably involved. Investigating...
